\name{OTRselect-package}
\alias{OTRselect-package}
\alias{OTRselect}
\docType{package}
\title{
\packageTitle{OTRselect}
}
\description{
\packageDescription{OTRselect}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{OTRselect}
\packageIndices{OTRselect}
Function \code{censored} performs variable selection for censored continuous response. Function \code{uncensored} performs variable selection for uncensored continuous response. Function \code{Qhat} estimates the restricted mean response given a treatment regime for censored data or the mean response given a treatment regime for uncensored data.
}
\author{
\packageAuthor{OTRselect}

Maintainer: \packageMaintainer{OTRselect}
}
\references{
Lu, W., Zhang, H. H., and Zeng. D. 
(2013). 
Variable selection for optimal treatment decision. 
Statistical Methods in Medical Research, 22, 493--504. 
PMCID: PMC3303960.

Geng, Y., Lu, W., and Zhang, H. H.
(2015).
On optimal treatment regimes selection for mean survival time.
Statistics in Medicine, 34, 1169--1184.
PMCID: PMC4355217.
}
\keyword{ package }

